reduce server-side fs reads, refs #332
