+++ src/server/utils/misc.js
@@ -1,3 +1,5 @@
+import path from 'path';
+import fs from 'fs';
 import { esc } from 'db/models';
 
 const emailRegex = /^([^\x00-\x20\x22\x28\x29\x2c\x2e\x3a-\x3c\x3e\x40\x5b-\x5d\x7f-\xff]+|\x22([^\x0d\x22\x5c\x80-\xff]|\x5c[\x00-\x7f])*\x22)(\x2e([^\x00-\x20\x22\x28\x29\x2c\x2e\x3a-\x3c\x3e\x40\x5b-\x5d\x7f-\xff]+|\x22([^\x0d\x22\x5c\x80-\xff]|\x5c[\x00-\x7f])*\x22))*\x40([^\x00-\x20\x22\x28\x29\x2c\x2e\x3a-\x3c\x3e\x40\x5b-\x5d\x7f-\xff]+|\x5b([^\x0d\x5b-\x5d\x80-\xff]|\x5c[\x00-\x7f])*\x5d)(\x2e([^\x00-\x20\x22\x28\x29\x2c\x2e\x3a-\x3c\x3e\x40\x5b-\x5d\x7f-\xff]+|\x5b([^\x0d\x5b-\x5d\x80-\xff]|\x5c[\x00-\x7f])*\x5d))*$/;
@@ -57,9 +59,22 @@ function checkCSRF(ctx, csrf) {
     return true;
 }
 
+function getSupportedLocales() {
+    const locales = [];
+    const files = fs.readdirSync(
+        path.join(__dirname, '../../..', 'src/app/locales')
+    );
+    for (const filename of files) {
+        const match_res = filename.match(/(\w+)\.json?$/);
+        if (match_res) locales.push(match_res[1]);
+    }
+    return locales;
+}
+
 module.exports = {
     emailRegex,
     getRemoteIp,
     rateLimitReq,
     checkCSRF,
+    getSupportedLocales,
 };
